requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
click==8.2.1
dateparser==1.2.1
distro==1.9.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.6
openai==1.90.0
pandas==2.3.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
rich==14.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typer==0.12.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
langgraph
